updates meme index page to remove descriptions and titles, bug fix prevents a user from editing another user's meme if they add edit to url path

+++ controllers/memeControllers.js
@@ -65,6 +65,8 @@ const memeController = {
         }
     },
     edit: async (req, res) => {
+        const findUser = await User.findOne({_id:req.params.id});
+        if (findUser._id.toString() === req.session.userId.toString()){
             try {
                 const foundMeme = await Meme.findById(req.params.id)
                 res.render('memes/edit.ejs', {
@@ -74,6 +76,10 @@ const memeController = {
             } catch(err) {
                 res.send(err)
             }
+        } else {
+            req.session.message = 'Sorry, you do not have permission to edit this meme.';
+            res.redirect(`/users/${findUser._id}`);
+        }
     },
     update: async (req, res) => {
         if(req.body.isVideo && req.body.isVideo=== 'on') {
@@ -91,11 +97,17 @@ const memeController = {
         }
     },
     delete: async(req,res) => {
-        try {
-          const foundMeme = await Meme.findByIdAndRemove(req.params.id);
-          res.redirect('/memes')
-        } catch(err) {
-          res.send(err)
+        const findUser = await User.findOne({_id:req.params.id});
+        if (findUser._id.toString() === req.session.userId.toString()){
+            try {
+            const foundMeme = await Meme.findByIdAndRemove(req.params.id);
+            res.redirect('/memes')
+            } catch(err) {
+            res.send(err)
+            }
+        } else {
+            req.session.message = 'Sorry, you do not have permission to delete this meme.';
+            res.redirect(`/users/${findUser._id}`);
         }
     }
 }
